src/main/java/com/fastcampus/projectborad/dto/security/KakaoOAuth2Response.java
package com.fastcampus.projectborad.dto.security;

import java.time.LocalDateTime;
import java.time.ZoneId;
import java.time.format.DateTimeFormatter;
import java.util.Map;

public record KakaoOAuth2Response(
        Long id,
        Boolean hasSignedUp,
        LocalDateTime connectedAt,
        Map<String, Object> properties,
        KakaoAccount kakaoAccount
) {
    public record KakaoAccount(
            Boolean profileNicknameNeedsAgreement,
            Profile profile,
            Boolean hasEmail,
            Boolean emailNeedsAgreement,
            Boolean isEmailValid,
            Boolean isEmailVerified,
            String email
    ) {

        public record Profile(String nickname) {
            public static Profile from(Map<String, Object> attributes) {
                return new Profile(String.valueOf(attributes.get("nickname")));
            }
        }

        public static KakaoAccount from(Map<String, Object> attributes) {
            return new KakaoAccount(
                    Boolean.valueOf(String.valueOf(attributes.get("profile_nickname_needs_agreement"))),
                    Profile.from((Map<String, Object>) attributes.get("profile")),
                    Boolean.valueOf(String.valueOf(attributes.get(""))),
                    Boolean.valueOf(String.valueOf(attributes.get("email_needs_agreement"))),
                    Boolean.valueOf(String.valueOf(attributes.get("is_email_valid"))),
                    Boolean.valueOf(String.valueOf(attributes.get("is_email_verified"))),
                    String.valueOf(attributes.get("email"))
            );
        }
    }

    public static KakaoOAuth2Response from(Map<String, Object> attribute) {
        return new KakaoOAuth2Response(
                Long.valueOf(String.valueOf(attribute.get("id"))),
                Boolean.valueOf(String.valueOf(attribute.get("has_signed_up"))),
                LocalDateTime.parse(
                        String.valueOf(attribute.get("connected_at")),
                        DateTimeFormatter.ISO_INSTANT.withZone(ZoneId.systemDefault())),
                (Map<String, Object>) attribute.get("properties"),
                KakaoAccount.from((Map<String, Object>) attribute.get("kakao_account"))
        );
    }
}
